src/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Flask==2.0.2
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-bigquery==2.30.1
google-cloud-bigquery-storage==2.10.1
google-cloud-core==2.2.1
google-cloud-logging==2.7.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
libcst==0.3.23
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.4
packaging==21.3
pandas==1.3.4
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.8
six==1.16.0
typing-inspect==0.7.1
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
